BDS: fix panel outputmode mismatch on TV [1/2]
PD#SWPL-134619

Problem:
The inability to display the panel logo during
the boot process arises from the HDMI TX mode
being configured for the panel.

Solution:
Currently, the HWC ModePolicy exclusively supports OTT.
To address this, it is necessary to disable it for BDS
and continue using systemControl.

Verify:
A311D2

Change-Id: Ia16b34720cbf2124bdddc29ccca6f136591cb7d7
Signed-off-by: Cheng Wang2 <cheng.wang2@amlogic.com>