audio: set max volume to 100 for all stream [2/3]
PD#SWPL-177225

Problem:
testVolumeKey_whileSessionAlive 1fail

Solution:
1. set max volume to 100 for all stream
2. ro.config.media_vol_steps will override the maximum volume
of music, and keep all other volumes at the maximum volume of 100.
In this way, when the music volume is converted to the volume
of other streams, there will be no precision loss caused by mute.
Because the alias of all streams is music.

Verify:
ohm/calla + U

Change-Id: I3b1d260e203b092d045baf067fd03ca59d838847
Signed-off-by: hongchao.yin <hongchao.yin@amlogic.com>